requirements.txt
requests
pyfiglet
fake-useragent
beautifulsoup4
geoip2
python-nmap
tldextract
lxml
pymongo
virustotal_python
dnspython
pysocks
asyncio
aiohttp
chardet
uvloop
urllib3